First cut at reworking the queue management for SDI multi-channel audio
This patch removes the cut/pasted FIFO implementation from libav
and replaces it with the use of VLC native block_Fifo queues.  We
now queue the blocks on arrival rather than copying the content
out of them, which preserves the timestamp information.

We also reworked the algorithm to monitor for the different queues
being heavily out-of-sync from a timestamp perspective.  This
allows us to throw away samples older than the PTS we are currently
rendering in the first channel (we're clocking everything against
the first channel).  It also allows us to hold off on dequeueing
buffers which are far in advance of the first channel PTS, putting
out silence in it's place (and more importantly, not rendering out
the audio for that channel way too early).

There is still some debug in this patch, and it's likely the whole
branch will get squashed onto a single commit before merging back
to trunk.